requirements.txt
numpy==1.23.5
opencv_python==4.7.0.72
pycryptodome==3.17
